Initial commit of query distance calculation.
We now compute and return the distances of locations from the given query point.

We're successfully converting between different units (e.g. we're returning miles or km based on the submitted user query).

Looks like we'll need some cleverness to actually get the correct system as it's only stored in the javax Quantity as a conversion.