Add distance calculation to near query bundle (#62)
* Initial commit of query distance calculation.

We now compute and return the distances of locations from the given query point.

We're successfully converting between different units (e.g. we're returning miles or km based on the submitted user query).

Looks like we'll need some cleverness to actually get the correct system as it's only stored in the javax Quantity as a conversion.

* LocationEntity now handles distance units correctly.

Updated VaccineLocation profile to account for the fact that we might return a distance extension.

* Reorder extension definition to fix failing test.

I know order was important, I just didn't know which order mattered.