Centralize Meta generation into base entity
Rather than generating the Meta component in each entity, we do it centrally.
There's a bit of a conflation of concerns because our Telecom and Identifier Entities are also included in this, but we never all `generateMeta` from those classes, so it should be fine.
We also just return an empty string, so at least the error will be noticeable.