Added broken support for contained resources to Slot Provider
Tried to add the ability to include the schedule for a given slot, but it doesn't work for some reason.

As a fallback, we're now returning an additional extension with a `forLocation` system to point to the transitive location.

It's a hack and it's not actually correct, but it's working.

Also fixed failing tests due to missing CORS configuration.